+++ core/HoistService.ts
@@ -31,7 +31,7 @@ import {apiDeprecated} from '@xh/hoist/utils/js';
  * so desired. And while components should typically source their state from backing models, they
  * can also read and react to service state and call service APIs.
  *
- * @mcpRole base class for all application services
+ * @mcpHint base class for all application services
  */
 export class HoistService extends HoistBase implements Loadable {
     // Internal State