+++ core/XH.ts
@@ -87,7 +87,7 @@ declare const xhIsDevelopmentMode: boolean;
  *
  * Available via import as `XH` - also installed as `window.XH` for troubleshooting purposes.
  *
- * @mcpRole singleton (XH) providing global framework services
+ * @mcpHint singleton (XH) providing global framework services
  */
 export class XHApi {
     /** Unique id for this loaded instance of the app.  Unique for every refresh of document. */